Create a trimesh object

These functions helps in creating trimesh objects. A trimesh is a subclass of mesh3d as defined in the rgl package. It is exclusively using triangles (unsurprisingly) and adds two additional elements: it_info which is a data.frame with information about the triangles in the mesh, and vb_info which is a data.frame with information about the vertices. Since trimesh is all about vector graphics it doesn't concern itself with material, texcoords, and normals — converting from a mesh3d will loose that information. trimesh objects can be constructed from either mesh3d object or data frames containing an x, y, and z column (remaining columns will go to the it_info element), or potentially from raw data giving vertice coordinates and triangle vertex indexes.

new_trimesh(vertices, triangles, vertex_info = NULL, triangle_info = NULL)

is_trimesh(mesh)

as_trimesh(mesh, ...)

Arguments

vertices
A 4-row matrix giving the coordinates of vertices in 3D space (last row is filled with 1's by convention). Each vertex correspond to a column in the matrix

triangles
A 3-row matrix giving the indexes of the three vertices that corresponds to the triangle. Each triangle corresponds to a column in the matrix

vertex_info
A data.frame giving additional information about each vertex in the vertices matrix.

triangle_info
A data.frame giving addtional information about each triangle given in the triangles matrix

mesh
A trimesh or an object convertible to one

...
ignored